+++ Sources/Taps/Tapes/TodoTaps.swift
@@ -1,10 +1,10 @@
 import RxSwift
 import TestHarness
 
-public struct TodoTaps: OfferingTaps {
-  public let testCaseObserver: AnyObserver<TestCase>
+private struct TodoTaps: OfferingTaps {
+  let testCaseObserver: AnyObserver<TestCase>
 
-  fileprivate init(offering taps: OfferingTaps, message: String?) {
+  init(offering taps: OfferingTaps, message: String?) {
     testCaseObserver = taps.testCaseObserver.mapObserver { testCase in
       guard testCase.directive == nil else { return testCase }
       var testCase = testCase
@@ -15,11 +15,15 @@ public struct TodoTaps: OfferingTaps {
 }
 
 public extension OfferingTaps {
-  public func todo(_ message: String? = nil) -> TodoTaps {
+  /// All tests defined inside the use of directive will be marked as todo.
+  /// If they fail, the test is marked as not ok, but as passing.
+  public func todo(_ message: String) -> OfferingTaps {
     return TodoTaps(offering: self, message: message)
   }
 
-  public var todo: TodoTaps {
+  /// All tests defined inside the use of directive will be marked as todo.
+  /// If they fail, the test is marked as not ok, but as passing.
+  public var todo: OfferingTaps {
     return TodoTaps(offering: self, message: nil)
   }
 }